Extract Range and Date in seprate component
