+++ components/FilterExpense/index.jsx
@@ -1,33 +1,20 @@
+import { useState } from "react";
+import { Icon } from "@iconify/react";
 import {
   CollapseContainer,
   CollapsiblePanelContainer,
   FilterButton,
   Select,
-  DatePickerWrapper,
-  CustomDatePicker,
-  RangeButton,
-  RangeSlider,
-  SelectedRangeValue,
-  RangeSliderContainer,
 } from "@/design-system/StyledFilterExpense";
-import { useState } from "react";
-import { Icon } from "@iconify/react";
-import "react-datepicker/dist/react-datepicker.css";
+
+import AmountRange from "../RangeSlider";
+import DatePicker from "../DatePicker";
 
 function FilterExpense({ selectedCategory, categoryNames, onCategoryFilter }) {
   const [collapsed, setCollapsed] = useState(true);
-  const [startDate, setStartDate] = useState(null);
-  const [endDate, setEndDate] = useState(null);
-
-  const [collapsedRange, setCollapsedRange] = useState(true);
-  const [rangeValue, setRangeValue] = useState(0);
 
   const toggleCollapse = () => {
     setCollapsed(!collapsed);
-  };
-
-  const toggleCollapseRange = () => {
-    setCollapsedRange(!collapsedRange);
   };
 
   return (
@@ -50,50 +37,9 @@ function FilterExpense({ selectedCategory, categoryNames, onCategoryFilter }) {
             ))}
           </Select>
 
-          {/* ////////Range////////// */}
-            <RangeSliderContainer>
-          <RangeButton onClick={toggleCollapseRange}>
-            Select Amount Range
-          </RangeButton>
-          <CollapsiblePanelContainer collapsed={collapsedRange}>
-              <RangeSlider
-                type="range"
-                min="0"
-                max="10000"
-                step="50"
-                value={rangeValue}
-                onChange={(event) => {
-                  setRangeValue(event.target.value);
-                }}
-              />
-              <SelectedRangeValue>{rangeValue} €</SelectedRangeValue>
-          </CollapsiblePanelContainer>
-            </RangeSliderContainer>
-          {/* ///////////////////////// */}
-
-          {/* ////////Date////////// */}
-          <DatePickerWrapper>
-          <span> from </span>
-            <CustomDatePicker
-              placeholderText="Start Date"
-              selected={startDate}
-              onChange={(date) => setStartDate(date)}
-              selectsStart
-              startDate={startDate}
-              endDate={endDate}
-            />
-            <span> to </span>
-            <CustomDatePicker
-              placeholderText="End Date"
-              selected={endDate}
-              onChange={(date) => setEndDate(date)}
-              selectsEnd
-              startDate={startDate}
-              endDate={endDate}
-              minDate={startDate}
-            />
-          </DatePickerWrapper>
-          {/* ///////////////////////// */}
+          <AmountRange />
+          <DatePicker />
+          
         </CollapsiblePanelContainer>
       </CollapseContainer>
     </>
